+++ gulpfile.js
@@ -11,6 +11,7 @@ var rename = require('gulp-rename');
 var tslint = require('gulp-tslint');
 var sassLint = require('gulp-sass-lint');
 var ts = require('gulp-typescript');
+var replace = require('gulp-replace');
 
 // PostCSS plugins
 var postcssSVG = require('postcss-svg');
@@ -29,6 +30,8 @@ var del = require('del');
 var runSequence = require('run-sequence');
 var browserSync = require('browser-sync');
 var merge = require('merge2');
+var nativeTslint = require('tslint');
+var npmPackage = require('./package.json');
 
 var paths = {
   source: {
@@ -61,40 +64,13 @@ gulp.task('clean', del.bind(null, [paths.target.html]));
 
 // TypeScript linting
 gulp.task('lint-ts', function() {
+  // The program is required for type checking rules to work: https://palantir.github.io/tslint/usage/type-checking/
+  var program = nativeTslint.Linter.createProgram("./tsconfig.json");
+
   return gulp.src(paths.source.ts)
   .pipe(tslint({
     formatter: 'verbose',
-    configuration: {
-      rules: {
-        'class-name': true,
-        'comment-format': [true, 'check-space'],
-        'indent': [true, 'spaces'],
-        'no-duplicate-variable': true,
-        'no-eval': true,
-        'no-internal-module': true,
-        'no-trailing-whitespace': true,
-        'no-var-keyword': false,
-        'one-line': [true, 'check-open-brace', 'check-whitespace'],
-        'quotemark': [true, 'single'],
-        'semicolon': false,
-        'triple-equals': [true, 'allow-null-check'],
-        'typedef-whitespace': [true, {
-          'call-signature': 'nospace',
-          'index-signature': 'nospace',
-          'parameter': 'nospace',
-          'property-declaration': 'nospace',
-          'variable-declaration': 'nospace'
-        }],
-        'variable-name': [true, 'ban-keywords'],
-        'whitespace': [true,
-          'check-branch',
-          'check-decl',
-          'check-operator',
-          'check-separator',
-          'check-type'
-        ]
-      }
-    }
+    program: program,
   }))
   .pipe(tslint.report({
     // Print just the number of errors (instead of printing all errors again)
@@ -140,7 +116,8 @@ gulp.task('browserify', function() {
   // Compile output JS file
   var stream = browserifyBundle
   .pipe(source('bitmovinplayer-ui.js'))
-  .pipe(buffer())
+  .pipe(replace('{{VERSION}}', npmPackage.version))
+  .pipe(buffer()) // required for production/sourcemaps
   .pipe(gulp.dest(paths.target.js));
 
   if (production) {