+++ src/ts/components/hugereplaybutton.ts
@@ -1,7 +1,6 @@
 import {ButtonConfig, Button} from './button';
 import {DOM} from '../dom';
 import {UIInstanceManager} from '../uimanager';
-import PlayerEvent = bitmovin.PlayerAPI.PlayerEvent;
 
 /**
  * A button to play/replay a video.
@@ -13,7 +12,7 @@ export class HugeReplayButton extends Button<ButtonConfig> {
 
     this.config = this.mergeConfig(config, {
       cssClass: 'ui-hugereplaybutton',
-      text: 'Replay'
+      text: 'Replay',
     }, this.config);
   }
 
@@ -34,7 +33,7 @@ export class HugeReplayButton extends Button<ButtonConfig> {
     // can cover the whole video player are and scaling would extend it beyond. By adding an inner element, confined
     // to the size if the image, it can scale inside the player without overshooting.
     buttonElement.append(new DOM('div', {
-      'class': this.prefixCss('image')
+      'class': this.prefixCss('image'),
     }));
 
     return buttonElement;