+++ src/ts/player.d.ts
@@ -35,10 +35,11 @@ declare namespace bitmovin {
         EASE_IN: PlayerAPI.VR.TransitionTimingType,
         EASE_OUT: PlayerAPI.VR.TransitionTimingType,
         EASE_IN_OUT: PlayerAPI.VR.TransitionTimingType,
-      }
-    }
+      },
+    };
   }
 
+  // tslint:disable-next-line:no-unused-variable
   const player: PlayerStatic;
 
   /**
@@ -516,8 +517,8 @@ declare namespace bitmovin {
         EASE_IN: PlayerAPI.VR.TransitionTimingType,
         EASE_OUT: PlayerAPI.VR.TransitionTimingType,
         EASE_IN_OUT: PlayerAPI.VR.TransitionTimingType,
-      }
-    }
+      },
+    };
 
     vr: PlayerAPI.PlayerVRAPI;
   }
@@ -937,7 +938,7 @@ declare namespace bitmovin {
           /**
            *  Two equirectangular videos for 3D content in side-by-side position.
            */
-        SBS
+        SBS,
       }
 
       enum State {